+++ return_invoice_bill/doc/RELEASE_NOTES.md
@@ -3,4 +3,11 @@
 #### 27.02.2024
 #### Version 16.0.1.0.0
 #### ADD
-- Initial commit for Return Invoices And Bills     
+- Initial commit for Automated Credit/Debit Note Generation from Return Picking     
+
+#### 05.08.2024
+#### Version 16.0.1.1.0
+#### UPDT
+
+- Updated the return invoice line based on the return quantity.
+- Fixed the issues while changing the language.